requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0

# Vector Database
chromadb>=0.4.0

# AI/ML
openai>=1.0.0
rank-bm25>=0.2.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
